03/12 20:55
feat: 결제 할인 추가

자신에게 맞는 할인율을 입력받아 최종계산에 적용하는 동작 완성 제네릭을 통해 가격의 여러가지 타입을 받아옴

+++ src/Menu.java
@@ -62,15 +62,18 @@ public class Menu {
     public MenuItem printMenuItem(int input){
            return tmpMenuItems.get(input-1);
     }
+
+    // 장바구니에 담겨있는 음식들 출력
     public void printShoppingCartItems(){
         for(MenuItem i : cartMenuItems){
             System.out.println(i);
         }
     }
+    //장바구니에 담겨있는 음식들의 값의 합 계산
     public Double shoppingCartItemsPrice(){
-        Double sum = 0.0;
+        double sum=0.0;
         for(MenuItem i : cartMenuItems){
-            sum += i.getPrice();
+            sum += i.getPrice().doubleValue();
         }
         return sum;
     }
